agent/requirements.txt
livekit-agents>=0.12.1
livekit-plugins-openai>=0.10.9
livekit-plugins-azure
livekit-plugins-silero>=0.7.4
livekit-plugins-turn_detector
python-dotenv~=1.0
